rename files and move logic
